creating a tasks module
